Search by Tx Hash
WIP: Search by Tx Hash

+++ service/src/main/java/net/tychecash/explorer/service/model/response/tx/Value.java
@@ -5,6 +5,7 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import java.math.BigInteger;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
@@ -15,19 +16,19 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class Value {
 
     @JsonProperty("amount")
-    private Integer amount;
+    private BigInteger amount;
     @JsonProperty("k_image")
     private String kImage;
     @JsonProperty("key_offsets")
     private List<Integer> keyOffsets = null;
 
     @JsonProperty("amount")
-    public Integer getAmount() {
+    public BigInteger getAmount() {
         return amount;
     }
 
     @JsonProperty("amount")
-    public void setAmount(Integer amount) {
+    public void setAmount(BigInteger amount) {
         this.amount = amount;
     }
